wlplan/tests/test-requirements.txt
--extra-index-url https://test.pypi.org/simple/  # for my pymimir fork
ipykernel==6.29.4
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
pymimir-dzc-fork==0.1.3
pytest==8.2.1
scikit-learn==1.5.0
